Implement pow (carrot) operator

+++ cppcalcpp/evaluator.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <cstdint>
 #include "exprtree.h"
-namespace evaluator
-{
-    int32_t evaluate(const exprtree* tree);
+namespace evaluator {
+    int64_t evaluate(const exprtree* tree);
 }
